// src/app/modules/order/models/point.model.ts
import { BaseModel } from 'app/models/BaseModel';
import { ContactModel } from 'app/models/contact.model';
import { CostDetail } from './cost-detail.model';
import { IIncident } from './point-incident.model';
import { ILocation } from 'app/modules/order/models/location.model';
import { PointStatus } from 'app/modules/order/constants/PointStatus';
import { PointType } from 'app/modules/order/constants/PointType';
import { ProductModel } from './product.model';
import { ServiceModel } from 'app/modules/price/models/service.model';

export class PointModel extends BaseModel {
    userIncidents?: IIncident[];
    images?: string[];
    note?: string;
    partnerNote?: string;
    signImage?: string;
    id?: string;
    distance?: number;
    expectedTime?: number;
    location?: ILocation;
    status?: PointStatus;
    deliveryAt?: number;
    services?: ServiceModel[];
    additionCost?: ServiceModel[];
    unexpectedIncidents?: IIncident[];
    finishedAt?: number;
    arrivedAt?: number;
    installedAt?: number;
    contact = new ContactModel();
    goodsCheckRequired = false;
    type: PointType;
    pickupAt: number;
    arrviedAt: number;
    productCodes: string;
    productSerialCodes: string;
    productStatus: string;
    productTypeId: string;
    requestType: ['consultation', 'home-service', 'store-service'];
    scheduleTime: number;
    brandId?: string;
    storeId?: string;
    userCost?: number;
    realCost?: number;
    costDetail: CostDetail[] = [];
    hostSignImage?: string;
    renterSignImage?: string;
    paid?: number;
    products: ProductModel[] = [];
    pickPointId?: string;
    externalCode?: string;
    laterReason?: string;
    scanStore?: boolean;
}